fix: refactor codebase
